escape arbitrary sentence text from clipboard
Running `ames.sh -c` copies the current text in the clipboard into
the sentence field in Anki. However, the current text is directly
substituted unescaped into the JSON request to AnkiConnect. If the
current text contains the '"' (double quote) or '\' (backslash)
characters, this can cause the JSON string itself to end early.

This commit fixes this issue by first escaping the
text by substituting '\' -> '\\' and then '"' -> '\"'.