more graceful exits on SIGINT and better errors
In the previous version, the `Pa_StartStream()` call to start
recording was placed within the SIGINT handling block. If a SIGINT
signal was recieved during the time that the stream was starting,
then the internal state of the stream could be damaged, causing
the subsequent cleanup call to `Pa_Terminate()` to hang.

This commit fixes the issue by moving `Pa_StartStream()`
to be the very last thing before entering SIGINT handling.
In addition, a `sig_check()` is added directly after `Pa_Initialize()`
to catch SIGINT after PortAudio initialization as well.

Error checking is more robust and thorough across the board.